==> notebooks/0_join_data.ipynb <==
import pandas as pd

As the data is a bit heavy the dataset is split in 2 csv.
Let's try joining it to work with only one csv.

--------------------------------

Damage Tables

dmg_demos_1 = pd.read_csv('../data/csgo/esea_master_dmg_demos.part1.csv')
dmg_demos_2 = pd.read_csv('../data/csgo/esea_master_dmg_demos.part2.csv')

dmg_demos = pd.concat([dmg_demos_1, dmg_demos_2])

dmg_demos.shape

dmg_demos.to_csv('../data/csgo/esea_master_dmg_demos.csv')

Drop some columns to reduce size

dmg_demos.drop(columns=['seconds', 'att_rank', 'vic_rank',
                        'att_pos_x', 'att_pos_y', 'vic_pos_x', 'vic_pos_y'], inplace=True)

dmg_demos.rename(columns={'Unnamed: 0':'index'}, inplace=True)

dmg_demos.to_csv('../data/csgo/esea_master_dmg_clean_demos.csv')

--------------
Granade Tables

nade_demos_1 = pd.read_csv('../data/csgo/esea_master_grenades_demos.part1.csv')
nade_demos_2 = pd.read_csv('../data/csgo/esea_master_grenades_demos.part2.csv')

print(nade_demos_1.shape, nade_demos_2.shape)

nade_demos = pd.concat([nade_demos_1, nade_demos_2])

nade_demos.shape

nade_demos.to_csv('../data/csgo/esea_master_grenades_demos.csv')

Drop some columns to reduce size

nade_demos.drop(columns=['seconds', 'vic_id', 'hp_dmg', 'arm_dmg', 'hitbox', 'att_rank', 'vic_rank',
                        'att_pos_x', 'att_pos_y', 'nade_land_x', 'nade_land_y', 'vic_pos_x', 'vic_pos_y'], inplace=True)

nade_demos.to_csv('../data/csgo/esea_master_grenades_clean_demos.csv')

---------------------
Kills Tables

kills_demos_1 = pd.read_csv('../data/csgo/esea_master_kills_demos.part1.csv')
kills_demos_2 = pd.read_csv('../data/csgo/esea_master_kills_demos.part2.csv')

print(kills_demos_1.shape, kills_demos_2.shape)

kills_demos = pd.concat([kills_demos_1, kills_demos_2])

kills_demos.shape

kills_demos.to_csv('../data/csgo/esea_master_kills_demos.csv')

---------------------------------------
Meta Tables

meta_demos_1 = pd.read_csv('../data/csgo/esea_meta_demos.part1.csv')
meta_demos_2 = pd.read_csv('../data/csgo/esea_meta_demos.part2.csv')

print(meta_demos_1.shape, meta_demos_2.shape)

meta_demos = pd.concat([meta_demos_1, meta_demos_2])

meta_demos.shape

meta_demos.to_csv('../data/csgo/esea_meta_demos.csv')
