==> notebooks/1_direct_estimation.ipynb <==
## **Two ways to estimate the team value**
### **1st Direct estimation from records**
### **2nd Supervised ML regression algorithm**

import pandas as pd
from sklearn.metrics import mean_squared_error

pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', 30)

### Weapon & Grenade Prices

wp_values = {'USP': 200, 'P2000': 200, 'Glock': 200, 'P250': 300, 'DualBarettas': 500, 'Tec9': 500, 'FiveSeven': 500, 'CZ': 500, 'Deagle': 700, 
             'Nova': 1200, 'Swag7': 1800, 'SawedOff': 1200, 'XM1014': 2000, 'M249': 5200, 'Negev': 4000,
             'Mac10': 1050, 'MP9': 1250, 'MP7': 1700, 'UMP': 1200, 'Bizon': 1400, 'P90': 2350, 
             'Gallil': 2000, 'Famas': 2250, 'Scout': 1700, 'AK47': 2700, 'M4A4': 3100, 'M4A1': 3100, 'SG556': 3000, 'AUG': 3300, 'G3SG1': 5000, 'Scar20': 5000, 'AWP': 4750, 
             'Decoy': 50, 'Flash': 200, 'Smoke': 300, 'HE': 300, 'Incendiary': 600, 'Molotov': 400, 
             'Knife': 0, 'Zeus': 200}

### Data

dmg_demos = pd.read_csv('../data/csgo/esea_master_dmg_clean_demos.csv')
nade_demos = pd.read_csv('../data/csgo/esea_master_grenades_clean_demos.csv')

## **1st Direct estimacion from records**

### Team Value

def plus_armor(team_value):
    """
    A simple way to add armor value to team value. 
    It could be more precise.
    """
    if team_value <= 1250:
        return team_value + 3250 # 5 chest armor (650$ each)
    elif team_value > 1250:
        return team_value + 4000 # 5 full armor (1000$ each)

def defuse_kit(team_value):
    """
    Add defuse kit values if the team value reaches a minimum value
    Only in CT data!
    """
    if team_value > 10000:
        return 5000 # 5 defuse kits (400$ each)
    else:
        return 0  

# Let's take first only one match

file = 'esea_match_13770997.dem'

# rounds = dmg_demos[dmg_demos['file'] == file]['round'].max()

# Change the way to select the round number. 
# There are some missing rounds so lets take literally the round, not the max to do a range

rounds = dmg_demos[dmg_demos['file'] == file]['round'].unique()

df_file = dmg_demos[dmg_demos['file'] == file]
df_file_nade = nade_demos[nade_demos['file'] == file]

lst_ct_val = []
lst_t_val = []

for rnd in rounds:
    
    # 1st WAPONS
    
    # ---------> Put Bomb and Unknown filters!
    
    df_round_ct = df_file[(df_file['round'] == rnd) & (df_file['wp_type'] != 'Grenade') & (df_file['att_side'] == 'CounterTerrorist')]
    df_round_t = df_file[(df_file['round'] == rnd) & (df_file['wp_type'] != 'Grenade') & (df_file['att_side'] == 'Terrorist')]
    
    wp_ct = df_round_ct[['att_id', 'wp']].drop_duplicates(subset='att_id')
    wp_t = df_round_t[['att_id', 'wp']].drop_duplicates(subset='att_id')
    # Drop_duplicates with subset to take the first record of each player. This way we reduce the error of taking a dropped gun
    
    wp_ct_val = sum([wp_values[weapon] for weapon in wp_ct['wp']])
    wp_t_val = sum([wp_values[weapon] for weapon in wp_t['wp']])
    # Value of the guns the team is carrying (CT & T)
    
    # 2nd GRENADES
    df_round_ct_nade = df_file_nade[(df_file_nade['round'] == rnd) & (df_file_nade['att_side'] == 'CounterTerrorist')]
    df_round_t_nade = df_file_nade[(df_file_nade['round'] == rnd) & (df_file_nade['att_side'] == 'Terrorist')]
    
    nade_ct = df_round_ct_nade[['att_id', 'nade']].drop_duplicates()
    nade_t = df_round_t_nade[['att_id', 'nade']].drop_duplicates()
    # Drop_duplicates to take all the nades of each player (we asume the error of taking dropped nades)
    
    nade_ct_val = sum([wp_values[nade] for nade in nade_ct['nade']])
    nade_t_val = sum([wp_values[nade] for nade in nade_t['nade']])
    # Value of the grenades the team is carrying (CT & T)
    
    # Join both values (Weapons & Grenades)
    ct_val = wp_ct_val + nade_ct_val
    t_val = wp_t_val + nade_t_val
    
    # Adding armor values
    ct_val = plus_armor(ct_val) + defuse_kit(ct_val)
    t_val = plus_armor(t_val)
    
    # Append the values to the lists lst_ct_val & lst_t_val
    lst_ct_val.append(ct_val)
    lst_t_val.append(t_val)

print(lst_ct_val)
print()
print(lst_t_val)

### Comparison

To make the comparison we take the team value from the DataFrame esea_meta_demos.csv

But first we have to filter and take only the games that have the full rounds records, as the are some games with missing rounds.

meta_demos = pd.read_csv('../data/csgo/esea_meta_demos.csv')

files_complete = [file for file in meta_demos['file'].unique() if\
                     meta_demos[meta_demos['file'] == file]['round'].max() == \
                     len(meta_demos[meta_demos['file'] == file])]

#### **Let's create a DataFrame with the team value of our generator and the real team value**

We will create lists with the data that will be the different columns

# LISTS

lst_file = []
lst_round = []
lst_ct_val_pred = []
lst_t_val_pred = []
lst_ct_val_real = []
lst_t_val_real = []
lst_round_type = []

## Attention! 4 hours to run this cell below

### This is the first attemp to obtain and create the DataFrame from the original data.
### There is an optimized version in 2_1_ml_preprocessingdata.ipynb

# LISTS

lst_file = []
lst_round = []
lst_ct_val_pred = []
lst_t_val_pred = []
lst_ct_val_real = []
lst_t_val_real = []
lst_round_type = []

# Create the loop around files and rounds
log = 0 # Log to see the progress

# Make some previous filtering to optimize the loop
flt_nade = dmg_demos['wp_type'] != 'Grenade'
flt_unknown = dmg_demos['wp_type'] != 'Unkown' # not well written in original data
flt_bomb = dmg_demos['wp'] != 'Bomb'

dmg_demos_flt = dmg_demos[flt_nade & flt_unknown & flt_bomb]

for file in files_complete:
    
    log += 1
    if log % 500 == 0: print(f'{log} / {len(files_complete)} ............')
    
#     df_file_dmg = dmg_demos[dmg_demos['file'] == file]
#     df_file_nade = nade_demos[nade_demos['file'] == file]
#     df_file_meta = meta_demos[meta_demos['file'] == file]
    
    df_file_dmg = dmg_demos_flt[dmg_demos['file'] == file]
    df_file_nade = nade_demos[nade_demos['file'] == file]
    df_file_meta = meta_demos[meta_demos['file'] == file]
    
    rounds = df_file_dmg['round'].unique()
    
    for rnd in rounds:
        # 1st WEAPONS
#         flt_nade = df_file_dmg['wp_type'] != 'Grenade'
#         flt_unknown = df_file_dmg['wp_type'] != 'Unkown' # not well written in original data
#         flt_bomb = df_file_dmg['wp'] != 'Bomb'
        
#         df_round_ct_dmg = df_file_dmg[(df_file_dmg['round'] == rnd) & flt_nade & flt_unknown & flt_bomb & (df_file_dmg['att_side'] == 'CounterTerrorist')]
#         df_round_t_dmg = df_file_dmg[(df_file_dmg['round'] == rnd) & flt_nade & flt_unknown & flt_bomb & (df_file_dmg['att_side'] == 'Terrorist')]
        df_round_ct_dmg = df_file_dmg[(df_file_dmg['round'] == rnd) & (df_file_dmg['att_side'] == 'CounterTerrorist')]
        df_round_t_dmg = df_file_dmg[(df_file_dmg['round'] == rnd) & (df_file_dmg['att_side'] == 'Terrorist')]
        
        wp_ct = df_round_ct_dmg[['att_id', 'wp']].drop_duplicates(subset='att_id')
        wp_t = df_round_t_dmg[['att_id', 'wp']].drop_duplicates(subset='att_id')
        # Drop_duplicates with subset to take the first record of each player. This way we reduce the error of taking a dropped gun

#         wp_ct_val = sum([wp_values[weapon] for weapon in wp_ct['wp']])
#         wp_t_val = sum([wp_values[weapon] for weapon in wp_t['wp']])    
        wp_ct_val_records = sum([wp_values[weapon] for weapon in wp_ct['wp']])
        n_wp = len(wp_ct['wp'])
        if n_wp > 0:
            wp_ct_val = wp_ct_val_records + ((wp_ct_val_records/n_wp) * (5 - n_wp))
        elif n_wp == 0:
            wp_ct_val = wp_ct_val_records

        wp_t_val_records = sum([wp_values[weapon] for weapon in wp_t['wp']])
        n_wp = len(wp_t['wp'])
        if n_wp > 0:
            wp_t_val = wp_t_val_records + ((wp_t_val_records/n_wp) * (5 - n_wp))    
        elif n_wp == 0:
            wp_t_val = wp_t_val_records
        # Value of the guns the team is carrying (CT & T)

        # 2nd GRENADES
        df_round_ct_nade = df_file_nade[(df_file_nade['round'] == rnd) & (df_file_nade['att_side'] == 'CounterTerrorist')]
        df_round_t_nade = df_file_nade[(df_file_nade['round'] == rnd) & (df_file_nade['att_side'] == 'Terrorist')]

        nade_ct = df_round_ct_nade[['att_id', 'nade']].drop_duplicates()
        nade_t = df_round_t_nade[['att_id', 'nade']].drop_duplicates()
        # Drop_duplicates to take all the nades of each player (we asume the error of taking dropped nades)

        nade_ct_val = sum([wp_values[nade] for nade in nade_ct['nade']])
        nade_t_val = sum([wp_values[nade] for nade in nade_t['nade']])
        # Value of the grenades the team is carrying (CT & T)

        # Join both values (Weapons & Grenades)
        ct_val = wp_ct_val + nade_ct_val
        t_val = wp_t_val + nade_t_val

        # Adding armor & defuse kit values
        ct_val = plus_armor(ct_val) + defuse_kit(ct_val)
        t_val = plus_armor(t_val)
        
        # Extracting real CT & T team value and round_type from meta_demos
        ct_val_real = df_file_meta[df_file_meta['round'] == rnd]['ct_eq_val'].values[0]
        t_val_real = df_file_meta[df_file_meta['round'] == rnd]['t_eq_val'].values[0]
        round_type = df_file_meta[df_file_meta['round'] == rnd]['round_type'].values[0]
        # .values to get only the column value, not also the index
        # [0] because .values return a 1 dimension array

        # Append the values to the previosly defined lists 
        lst_file.append(file)
        lst_round.append(rnd)
        lst_ct_val_pred.append(ct_val)
        lst_t_val_pred.append(t_val)
        lst_ct_val_real.append(ct_val_real)
        lst_t_val_real.append(t_val_real)
        lst_round_type.append(round_type)

df = pd.DataFrame([lst_file, lst_round, lst_ct_val_pred, lst_t_val_pred, lst_ct_val_real, lst_t_val_real, lst_round_type]).T
df.columns = ['file', 'round', 'ct_val_pred', 't_val_pred', 'ct_val_real', 't_val_real', 'round_type']

df.to_csv('../data/processed/estimated_team_value.csv', index=False) # Saved .csv with the full data comparison

### Load the DataFrame 

df = pd.read_csv('../data/processed/estimated_team_value.csv')

print(df.shape)
df.head()

### Differences estimated / real

df['ct_error'] = df['ct_val_real'] - df['ct_val_pred']
df['t_error'] = df['t_val_real'] - df['t_val_pred']

df.describe()

df.head()

print(f"CT Error: {mean_squared_error(y_pred=df['ct_val_pred'], y_true=df['ct_val_real'], squared=False)}")
print(f"T Error: {mean_squared_error(y_pred=df['t_val_pred'], y_true=df['t_val_real'], squared=False)}")
