==> notebooks/2_4_ml_regressor_lgbm_tuning.ipynb <==
## **2nd Supervised ML LightGBM regressor algorithm**

import pandas as pd

from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler, LabelEncoder
from sklearn.model_selection import train_test_split, RandomizedSearchCV
from sklearn.metrics import mean_squared_error, r2_score

from lightgbm import LGBMRegressor
from joblib import dump

pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', 30)

### Data

df = pd.read_csv('../data/processed/base_to_ml_predicted_team_value.csv')

### Preprocessing

Erase column 'round_type' because we cannot know this value from the live records

df.drop(['round_type'], axis=1, inplace=True)

df.tail()

We want to predict 'ct_val' and 't_val' so we will split the dataframe into 2 dataframes, one with CT data and other with T data

**ct_df** = ['file',
 'round',
 'wp_ct_val',
 'nade_ct_val',
 'ct_alive',
 't_alive',
 'ct_winner',
 'bomb_planted',
 'ct_cons_wins',
 't_cons_wins',
 'ct_val_real']


**t_df** = ['file',
 'round',
 'wp_t_val',
 'nade_t_val',
 'ct_alive',
 't_alive',
 'ct_winner',
 'bomb_planted',
 'ct_cons_wins',
 't_cons_wins',
 't_val_real']

# Two DataFrames, one per CT another for T

ct_df = df[['file', 'round', 'wp_ct_val', 'nade_ct_val', 'ct_alive', 't_alive', 'ct_winner', 'bomb_planted', 'ct_cons_wins', 't_cons_wins', 'ct_val_real']]
t_df = df[['file', 'round', 'wp_t_val', 'nade_t_val', 'ct_alive', 't_alive', 'ct_winner', 'bomb_planted', 'ct_cons_wins', 't_cons_wins', 't_val_real']]

display(ct_df.head())
t_df.tail()

# LabelEncoder to categorical feature -> file
le = LabelEncoder()

ct_df['file'] = le.fit_transform(ct_df.file)
t_df['file'] = le.fit_transform(t_df.file)

display(ct_df.head())
t_df.tail()

## Define features and target

# Define all the features

CT_FEATS = ['file', 'round', 'wp_ct_val', 'nade_ct_val', 'ct_alive', 't_alive', 'ct_winner', 'bomb_planted', 'ct_cons_wins', 't_cons_wins']
T_FEATS = ['file', 'round', 'wp_t_val', 'nade_t_val', 'ct_alive', 't_alive', 'ct_winner', 'bomb_planted', 'ct_cons_wins', 't_cons_wins']
CT_TARGET = 'ct_val_real'
T_TARGET = 't_val_real'

#Preprocessor

numeric_transformer = Pipeline(steps=[('scaler', StandardScaler())])

preprocessor_ct = ColumnTransformer(transformers=[('num', numeric_transformer, CT_FEATS)])
preprocessor_t = ColumnTransformer(transformers=[('num', numeric_transformer, T_FEATS)])

### **SPLITS**
We are going to use cross-validation, but we still hide 5% of the data to make a final prediction and be aware of possible overfitting.

ct_train, ct_test = train_test_split(ct_df, train_size=0.95)
t_train, t_test = train_test_split(t_df, train_size=0.95)

print(ct_train.shape, ct_test.shape)
print(t_train.shape, t_test.shape)

### **MODEL**

regressor = LGBMRegressor()

# ct_model = Pipeline(steps=[('ct_preprocessor', preprocessor_ct),
#                            ('regressor', regressor)])

t_model = Pipeline(steps=[('t_preprocessor', preprocessor_t),
                           ('regressor', regressor)])

# To obtain the best hyperparameters of our model, we test a wider range of values and then 
# we reduce until obtaining close but different values, in order to obtain the best results.

param_grid = {
    'regressor__num_leaves': [47,48,49,50,51,52],
    'regressor__n_estimators': [350,360,370,380,390,400,410,420,430],
    'regressor__min_data_in_leaf': [4,5,6,7,8,12,15],
    'regressor__max_depth': [5,6,7],
    'regressor__learning_rate': [0.02,0.03,0.04,0.045,0.05,0.055,0.06],
    'regressor__feature_fraction': [0.75,0.8,0.85,0.9,0.95,1],
    'regressor__bagging_frequency': [0.75,0.80,0.85,0.9,0.95,1],
    'regressor__bagging_fraction': [0.75,0.8,0.85,0.9,0.95,1],
}

#CT MODEL --------------------------------------

# grid_search = RandomizedSearchCV(ct_model, 
#                                  param_grid, 
#                                  cv=5, 
#                                  verbose=5, 
#                                  scoring='r2', 
#                                  n_jobs=-1,
#                                  n_iter=700)

# grid_search.fit(ct_train[CT_FEATS], ct_train[CT_TARGET])


# T MODEL ---------------------------------------

grid_search = RandomizedSearchCV(t_model, 
                                 param_grid, 
                                 cv=5, 
                                 verbose=5, 
                                 scoring='r2', 
                                 n_jobs=-1,
                                 n_iter=500)

grid_search.fit(t_train[T_FEATS], t_train[T_TARGET])

#Once we have trained the model with different hyperparameter values, 
#we see which parameters have obtained the best score:
grid_search.best_params_

# see the best score that we have trained:
grid_search.best_score_

### Best hyperparameter values for **ct_model**

{'regressor__num_leaves': 51,
 'regressor__n_estimators': 400,
 'regressor__min_data_in_leaf': 15,
 'regressor__max_depth': 7,
 'regressor__learning_rate': 0.06,
 'regressor__feature_fraction': 0.85,
 'regressor__bagging_frequency': 0.95,
 'regressor__bagging_fraction': 1}

### Best hyperparameter values for **t_model**

{'regressor__num_leaves': 51,
 'regressor__n_estimators': 400,
 'regressor__min_data_in_leaf': 15,
 'regressor__max_depth': 7,
 'regressor__learning_rate': 0.06,
 'regressor__feature_fraction': 0.85,
 'regressor__bagging_frequency': 0.95,
 'regressor__bagging_fraction': 1}

### CT Regressor

regressor = LGBMRegressor(num_leaves=51,
                          n_estimators=400,
                          min_data_in_leaf=15,
                          max_depth=7,
                          learning_rate=0.06,
                          feature_fraction=0.85,
                          bagging_frequency=0.95,
                          bagging_fraction=1)

### T Regressor

regressor = LGBMRegressor(num_leaves=51,
                          n_estimators=400,
                          min_data_in_leaf=15,
                          max_depth=7,
                          learning_rate=0.06,
                          feature_fraction=0.85,
                          bagging_frequency=0.95,
                          bagging_fraction=1)

### Models

ct_model = Pipeline(steps=[('ct_preprocessor', preprocessor_ct),
                           ('regressor', regressor)])

# t_model = Pipeline(steps=[('t_preprocessor', preprocessor_t),
#                            ('regressor', regressor)])

### Test hyperparameter values from gridsearch

ct_model.fit(ct_train[CT_FEATS], ct_train[CT_TARGET]);

# t_model.fit(t_train[T_FEATS], t_train[T_TARGET]);

### **CHECK PERFORMANCE**

y_ct_test = ct_model.predict(ct_test[CT_FEATS])
y_ct_train = ct_model.predict(ct_train[CT_FEATS])

# y_t_test = t_model.predict(t_test[T_FEATS])
# y_t_train = t_model.predict(t_train[T_FEATS])

# Mean_squared_error

print(f"CT test error: {mean_squared_error(y_pred=y_ct_test, y_true=ct_test[CT_TARGET], squared=False)}")
print(f"CT train error: {mean_squared_error(y_pred=y_ct_train, y_true=ct_train[CT_TARGET], squared=False)}")
# print()
# print(f"T test error: {mean_squared_error(y_pred=y_t_test, y_true=t_test[T_TARGET], squared=False)}")
# print(f"T train error: {mean_squared_error(y_pred=y_t_train, y_true=t_train[T_TARGET], squared=False)}")

# r2_score

print(f"CT test error: {r2_score(y_pred=y_ct_test, y_true=ct_test[CT_TARGET])}")
print(f"CT train error: {r2_score(y_pred=y_ct_train, y_true=ct_train[CT_TARGET])}")
print()
# print(f"T test error: {r2_score(y_pred=y_t_test, y_true=t_test[T_TARGET])}")
# print(f"T train error: {r2_score(y_pred=y_t_train, y_true=t_train[T_TARGET])}")

--------------------------------
### LGBM REGRESSOR
CT test error:______Base: 2710.808 || 0.93391______<font color='green'>Hyp. tuning: 2672.174 || 0.93586</font>

CT train error:______Base: 2698.860 || 0.93466______<font color='green'>Hyp. tuning: 2628.522 || 0.93788</font>

T test error:______Base: 2725.320 || 0.90388______<font color='green'>Hyp. tuning: 2704.993 || 0.90657</font>

T train error:______Base: 2728.544 || 0.90519______<font color='green'>Hyp. tuning: 2660.081 || 0.90937</font>

###  **Train model with full dataset**

#### **CT Model. Train & Save**

regressor = LGBMRegressor(num_leaves=51,
                          n_estimators=400,
                          min_data_in_leaf=15,
                          max_depth=7,
                          learning_rate=0.06,
                          feature_fraction=0.85,
                          bagging_frequency=0.95,
                          bagging_fraction=1)

ct_model = Pipeline(steps=[('ct_preprocessor', preprocessor_ct),
                           ('regressor', regressor)])

ct_model.fit(ct_df[CT_FEATS], ct_df[CT_TARGET])

ct_df_pred_val = ct_model.predict(ct_df[CT_FEATS])

dump(ct_model, '../models/ct_team_value.joblib') # Save model

ct_df['ct_val_pred'] = ct_df_pred_val

ct_df.to_csv('../data/results/ct_predicted_value.csv') # Export the CT DataFrame with the prediction column added

#### **T Model. Train & Save**

regressor = LGBMRegressor(num_leaves=51,
                          n_estimators=400,
                          min_data_in_leaf=15,
                          max_depth=7,
                          learning_rate=0.06,
                          feature_fraction=0.85,
                          bagging_frequency=0.95,
                          bagging_fraction=1)

t_model = Pipeline(steps=[('t_preprocessor', preprocessor_t),
                           ('regressor', regressor)])

t_model.fit(t_df[T_FEATS], t_df[T_TARGET])

t_df_pred_val = t_model.predict(t_df[T_FEATS])

dump(t_model, '../models/t_team_value.joblib') # Save model

t_df['t_val_pred'] = t_df_pred_val

t_df.to_csv('../data/results/t_predicted_value.csv') # Export the T DataFrame with the prediction column added
