==> notebooks/3_1_preprocessing_ml_reg_nxt_rnd_val.ipynb <==
## **Supervised ML regression algorithm to predict next round team value (CT & T)**
## **Preprocessing**

Input data: Obtained DataFrames from the previous prediction:
- ct_predicted_value
- t_predicted_value

Preprocces the data to create a new traget column, next round team value (**nxt_rnd_ct_val** & **nxt_rnd_t_val**)

import pandas as pd

pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', 30)

### Data

ct_df = pd.read_csv('../data/results/ct_predicted_value.csv')
t_df = pd.read_csv('../data/results/t_predicted_value.csv')

## Preprocessing

To train the model we need to know the team value for the next round. 

We will train with the predicted value for the current round and real value for the next round

ct_df.drop(columns=['Unnamed: 0'], inplace=True)
t_df.drop(columns=['Unnamed: 0'], inplace=True)

#### Create the columns **nxt_rnd_ct_val** & **nxt_rnd_t_val** filled with 0

ct_df['nxt_rnd_ct_val'] = 0
t_df['nxt_rnd_t_val'] = 0

files = ct_df['file'].unique()

ct_df.set_index(['file', 'round'], inplace=True)
t_df.set_index(['file', 'round'], inplace=True)

#### Filling the values with the real values of next rounds

%%time

a = 0

for file in files:
    a += 1
    if a%500 == 0:
        print(f'{a}/12185')
    for rnd in ct_df.loc[file].index[:-1]:
        ct_df.loc[(file, rnd), 'nxt_rnd_ct_val'] = ct_df.loc[(file, rnd+1), 'ct_val_pred']
        t_df.loc[(file, rnd), 'nxt_rnd_t_val'] = t_df.loc[(file, rnd+1), 't_val_pred']

### Reset index and save the DataFrames 

ct_df.reset_index(inplace=True)
t_df.reset_index(inplace=True)

ct_df.to_csv('../data/processed/3_base_predict_next_rnd_ct_val.csv', index=True)
t_df.to_csv('../data/processed/3_base_predict_next_rnd_t_val.csv', index=True)

### Load Data

ct_df = pd.read_csv('../data/processed/3_base_predict_next_rnd_ct_val.csv')
t_df = pd.read_csv('../data/processed/3_base_predict_next_rnd_t_val.csv')

display(ct_df.head())
display(t_df.head())
