==> notebooks/3_2_ml_regressor_next_round_value.ipynb <==
## **Supervised ML regression algorithm to predict next round team value (CT & T)**
## **Algorithm**

Input data: Obtained DataFrames from the previous prediction:
- ct_predicted_value
- t_predicted_value

Preprocces the data to create a new traget column, next round team value (**nxt_rnd_ct_val** & **nxt_rnd_t_val**)

import pandas as pd

from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler, LabelEncoder
from sklearn.model_selection import train_test_split, RandomizedSearchCV
from sklearn.metrics import mean_squared_error, r2_score

from sklearn.linear_model import Lasso

from lightgbm import LGBMRegressor
from joblib import dump, load

pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', 30)

### Load Data

ct_df = pd.read_csv('../data/processed/3_base_predict_next_rnd_ct_val.csv')
t_df = pd.read_csv('../data/processed/3_base_predict_next_rnd_t_val.csv')

display(ct_df.head())
display(t_df.head())

### Define Features and Target

CT_FEATS = ['file', 'round', 'wp_ct_val', 'nade_ct_val', 'ct_alive', 't_alive', 'ct_winner', 'bomb_planted', 'ct_cons_wins', 't_cons_wins', 'ct_val_pred']
T_FEATS = ['file', 'round', 'wp_t_val', 'nade_t_val', 'ct_alive', 't_alive', 'ct_winner', 'bomb_planted', 'ct_cons_wins', 't_cons_wins', 't_val_pred']

CT_TARGET = 'nxt_rnd_ct_val'
T_TARGET = 'nxt_rnd_t_val'

### Preprocessor

#Preprocessor

numeric_transformer = Pipeline(steps=[('scaler', StandardScaler())])

preprocessor_ct = ColumnTransformer(transformers=[('num', numeric_transformer, CT_FEATS)])
preprocessor_t = ColumnTransformer(transformers=[('num', numeric_transformer, T_FEATS)])

### **SPLITS**

ct_train, ct_test = train_test_split(ct_df)
t_train, t_test = train_test_split(t_df)

print(ct_train.shape, ct_test.shape)
print(t_train.shape, t_test.shape)

### **MODEL**

#### We take **LightGBM** as the regressor algorithm, like the previous prediction

regressor = LGBMRegressor(boosting_type='gbdt', 
                       bagging_freq=1, 
                       bagging_fraction = 0.9, 
                       n_estimators=100)

# regressor = Lasso()

# ct_model = Pipeline(steps=[('ct_preprocessor', preprocessor_ct),
#                            ('regressor', regressor)])

t_model = Pipeline(steps=[('t_preprocessor', preprocessor_t),
                           ('regressor', regressor)])

# ct_model.fit(ct_train[CT_FEATS], ct_train[CT_TARGET]);

t_model.fit(t_train[T_FEATS], t_train[T_TARGET]);

### **CHECK PERFORMANCE**

# y_ct_test = ct_model.predict(ct_test[CT_FEATS])
# y_ct_train = ct_model.predict(ct_train[CT_FEATS])

y_t_test = t_model.predict(t_test[T_FEATS])
y_t_train = t_model.predict(t_train[T_FEATS])

# Mean_squared_error

# print(f"CT test error: {mean_squared_error(y_pred=y_ct_test, y_true=ct_test[CT_TARGET], squared=False)}")
# print(f"CT train error: {mean_squared_error(y_pred=y_ct_train, y_true=ct_train[CT_TARGET], squared=False)}")
print()
print(f"T test error: {mean_squared_error(y_pred=y_t_test, y_true=t_test[T_TARGET], squared=False)}")
print(f"T train error: {mean_squared_error(y_pred=y_t_train, y_true=t_train[T_TARGET], squared=False)}")

# r2_score

# print(f"CT test error: {r2_score(y_pred=y_ct_test, y_true=ct_test[CT_TARGET])}")
# print(f"CT train error: {r2_score(y_pred=y_ct_train, y_true=ct_train[CT_TARGET])}")
print()
print(f"T test error: {r2_score(y_pred=y_t_test, y_true=t_test[T_TARGET])}")
print(f"T train error: {r2_score(y_pred=y_t_train, y_true=t_train[T_TARGET])}")

--------------------------------
### LGBM REGRESSOR
CT test error:_______8444.855 || <font color='red'>0.34713</font>

CT train error:______8392.133 || <font color='red'>0.35541</font>

T test error:________6927.958 || <font color='red'>0.36787</font>

T train error:_______6838.890 || <font color='red'>0.38061</font>

### We get a very unaccurate prediction with a very low r2 score

After so many attempts with new features, parameter tuning, and different regressors, we conclude to try to <font color='green'>**predict a range of team values**</font>, not the exact value
