==> notebooks/4_1_prepr_ml_clas_nxt_rnd_val.ipynb <==
## **Supervised ML classification algorithm to predict next round team value (CT & T)**
## **Analysis**

Analyze first the data to define the **number of bins** we are going to create and then predict.


import pandas as pd

pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', 50)

meta_demos = pd.read_csv('../data/csgo/esea_meta_demos.csv')

meta_demos.head()

meta_demos['round_type'].unique()

#### From esea_meta_demos.csv we have **round_type** with 5 different classes closely related to team values:

- PISTOL_ROUND
- ECO
- SEMI_ECO
- FORCE_BUY
- NORMAL

#### Let's analyze each type and try to figure out the possible range of team value for every class

meta_types = meta_demos[['round_type', 'ct_eq_val', 't_eq_val']]

display(meta_types.groupby('round_type').min())
display(meta_types.groupby('round_type').max())

Not reliable information because round_type could refer only to one team.

We can see it easily in ECO type.

First, let's define ECO rounds. An ECO round occurs when a team has not enough money to get great weapons to face against the other team. Usually, the members of the team buy weapons or utilities but saving around $2000\$$ for the next round. This way they have the chance the next round to win with a more equally confrontation

Understanding now the ECO round concept we see that the minimum value is $200\$$ and $400\$$ which makes sense. But looking at the maximum values we can see that those values are $42.050\$$ and $36.700\$$, extremely high for a team who cannot afford good equipment.

It means that the classification is made with the cheapest team value.

From the cheapest to the most expensive type, the classification is:
- ECO 
- SEMI_ECO
- FORCE_BUY
- NORMAL

PISTOL_ROUND is a special type that occurs the first round of each playable side (2 times per game)

## **Preprocessing**
#### Let's create a new column with the lowest team value for each round and analyze the round type.

meta_types['min_eq_val'] = meta_types[['ct_eq_val', 't_eq_val']].min(axis=1)

meta_types.head(10)

meta_types[['round_type', 'min_eq_val']].groupby('round_type').max()

We can see the **top value** of each category according to the given classification

### Data

Let's load the data obtained from the first prediction (2_4_ml_regressor_lgbm_tuning.ipynb):
- ct_predicted_value.csv
- t_predicted_value.csv

ct_df = pd.read_csv('../data/results/ct_predicted_value.csv')
t_df = pd.read_csv('../data/results/t_predicted_value.csv')

ct_df.drop(columns=['Unnamed: 0'], inplace=True)
t_df.drop(columns=['Unnamed: 0'], inplace=True)

display(ct_df.head())
display(t_df.head())

#### Incorporate **round_type** column to the dataframes ct_df & t_df

We can rescue this column from the csv **base_to_ml_predicted_team_value.csv** exported in 2_1_ml_preprocessingdata.ipynb

df_round_type = pd.read_csv('../data/processed/base_to_ml_predicted_team_value.csv')

df_round_type.head()

ct_df['round_type'] = df_round_type['round_type']
t_df['round_type'] = df_round_type['round_type']

display(ct_df.head())
display(t_df.head())

#### As mentioned before, this is not the best way to classify the rounds, because it is not correlated to each team.

Let's do a manual classification by given thresholds:

- ECO $\leqslant$ 14500
- 14500 $<$ SEMI_ECO $\leqslant$ 18400
- 18400 $<$ FORCE_BUY $\leqslant$ 22900
- 22900 $<$ NORMAL

**PISTOL_ROUND will be the only one class that remains as given**

def get_round_type(index, ct_df, t_df):
    ct_val = ct_df.loc[index, 'ct_val_real']
    if ct_df.loc[index, 'round_type'] == 'PISTOL_ROUND':
        ct_rnd_type = 'PISTOL_ROUND'
    elif ct_val <= 14500:
        ct_rnd_type = 'ECO'
    elif ct_val <= 18400 and ct_val > 14500:
        ct_rnd_type = 'SEMI_ECO'
    elif ct_val <= 22900 and ct_val > 18400:
        ct_rnd_type = 'FORCE_BUY'
    elif ct_val > 22900:
        ct_rnd_type = 'NORMAL'
        
    t_val = t_df.loc[index, 't_val_real']
    if t_df.loc[index, 'round_type'] == 'PISTOL_ROUND':
        t_rnd_type = 'PISTOL_ROUND'
    elif t_val <= 14500:
        t_rnd_type = 'ECO'
    elif t_val <= 18400 and t_val > 14500:
        t_rnd_type = 'SEMI_ECO'
    elif t_val <= 22900 and t_val > 18400:
        t_rnd_type = 'FORCE_BUY'
    elif t_val > 22900:
        t_rnd_type = 'NORMAL'
        
    return ct_rnd_type, t_rnd_type

lst_round_type = [get_round_type(index, ct_df, t_df) for index in ct_df.index]

df_round_type = pd.DataFrame(lst_round_type, columns=[['ct_round_type', 't_round_type']])
df_round_type.head()

ct_df['round_type'] = df_round_type[['ct_round_type']]
t_df['round_type'] = df_round_type[['t_round_type']]

display(ct_df.head())
display(t_df.head())

#### From the experience of playing the game for a very long time, I do not agree with the results obtained from the applied threshold 

Let's define a new classification with the next parameters:
- ECO
- MEDIUM (Minimum threshold value: Armor + cheapest SMG + Flash $\rightarrow$ 1900x5 =  9.500)
- FULL (minimum threshold value: Armor-helmet + AK47 + Grenade + Flash $\rightarrow$ 4200x5 = 21.000)

def get_round_type(index, ct_df, t_df):
    ct_val = ct_df.loc[index, 'ct_val_real']
    if ct_df.loc[index, 'round_type'] == 'PISTOL_ROUND':
        ct_rnd_type = 'PISTOL_ROUND'
    elif ct_val <= 9500:
        ct_rnd_type = 'ECO'
    elif ct_val <= 21000 and ct_val > 9500:
        ct_rnd_type = 'MEDIUM'
    elif ct_val > 21000:
        ct_rnd_type = 'FULL'
        
    t_val = t_df.loc[index, 't_val_real']
    if t_df.loc[index, 'round_type'] == 'PISTOL_ROUND':
        t_rnd_type = 'PISTOL_ROUND'
    elif t_val <= 9500:
        t_rnd_type = 'ECO'
    elif t_val <= 21000 and t_val > 9500:
        t_rnd_type = 'MEDIUM'
    elif t_val > 21000:
        t_rnd_type = 'FULL'
        
    return ct_rnd_type, t_rnd_type

lst_round_type = [get_round_type(index, ct_df, t_df) for index in ct_df.index]

df_round_type = pd.DataFrame(lst_round_type, columns=[['ct_round_type', 't_round_type']])
df_round_type.head()

ct_df['round_type'] = df_round_type[['ct_round_type']]
t_df['round_type'] = df_round_type[['t_round_type']]

display(ct_df.head())
display(t_df.head())

### Let's create the column to predict, the **next_round_type**

%%time

# This code does not look fast, but it is 3 times faster than other much more pretty and sort that 
# I did for another similar situation

files = ct_df['file'].unique()

nxt_rnd_type_ct = []
nxt_rnd_type_t = []

log = 0

for file in files:

    ct_df_file = ct_df[ct_df['file'] == file]
    t_df_file = t_df[t_df['file'] == file]
    
    rounds = ct_df_file['round'].unique()
    
    if log%1000 == 0:
        print(f'{log}/12185')    
    log += 1

    
    for rnd in rounds:
        if rnd == rounds[-1]:
            nxt_rnd_type_ct.append('LAST')
            nxt_rnd_type_t.append('LAST')
        else:
            nxt_rnd_type_ct.append(ct_df_file[ct_df_file['round'] == rnd + 1]['round_type'].values[0])
            nxt_rnd_type_t.append(t_df_file[t_df_file['round'] == rnd + 1]['round_type'].values[0])
            
ct_df['nxt_rnd_type'] = nxt_rnd_type_ct
t_df['nxt_rnd_type'] = nxt_rnd_type_t

display(ct_df.head())
display(t_df.head())

#### Last step before going into the ml classification algorithm, let's remove the real value columns, and then save the dataframes to make our predictions

ct_df.drop(columns=['ct_val_real'], inplace=True)
t_df.drop(columns=['t_val_real'], inplace=True)

ct_df.to_csv('../data/processed/4_base_predict_next_rnd_ct_type.csv', index=False)
t_df.to_csv('../data/processed/4_base_predict_next_rnd_t_type.csv', index=False)
