==> notebooks/4_3_ml_clas_nxt_rnd_val.ipynb <==
## **Supervised ML classification algorithm to predict next round team value (CT & T)**
## **Algorithm**

import pandas as pd

from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split, RandomizedSearchCV
from sklearn.metrics import classification_report

from lightgbm import LGBMClassifier
from joblib import dump

pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', 30)

### Data

ct_df = pd.read_csv('../data/processed/4_base_predict_next_rnd_ct_type.csv')
t_df = pd.read_csv('../data/processed/4_base_predict_next_rnd_t_type.csv')

display(ct_df.head())
display(t_df.head())

### Encoding

We use a manual encoding to be sure of being the same values for round_type and nxt_rnd_type.

If we use LabelEncoder() it could be possible to have different values for the same label.

ct_df['round_type'].unique()

round_type_dic = {'PISTOL_ROUND':0, 'ECO':1, 'MEDIUM':2, 'FULL':3, 'LAST':4}

ct_df['round_type'] = ct_df['round_type'].apply(lambda x: round_type_dic[x])
ct_df['nxt_rnd_type'] = ct_df['nxt_rnd_type'].apply(lambda x: round_type_dic[x])

t_df['round_type'] = t_df['round_type'].apply(lambda x: round_type_dic[x])
t_df['nxt_rnd_type'] = t_df['nxt_rnd_type'].apply(lambda x: round_type_dic[x])

display(ct_df.head())
display(t_df.head())

## Define features and target

CT_FEATS = ['file', 'round', 'wp_ct_val', 'nade_ct_val', 'ct_alive', 't_alive', 'ct_winner', 'bomb_planted', 'ct_cons_wins', 't_cons_wins', 'ct_val_pred', 'round_type']
T_FEATS = ['file', 'round', 'wp_t_val', 'nade_t_val', 'ct_alive', 't_alive', 'ct_winner', 'bomb_planted', 'ct_cons_wins', 't_cons_wins', 't_val_pred', 'round_type']
TARGET = 'nxt_rnd_type'

#Preprocessor

numeric_transformer = Pipeline(steps=[('scaler', StandardScaler())])

preprocessor_ct = ColumnTransformer(transformers=[('num', numeric_transformer, CT_FEATS)])
preprocessor_t = ColumnTransformer(transformers=[('num', numeric_transformer, T_FEATS)])

### **SPLITS**

# X_train, X_test, y_train, y_test = train_test_split(ct_df[CT_FEATS], ct_df[TARGET]) # CT Team

X_train, X_test, y_train, y_test = train_test_split(t_df[T_FEATS], t_df[TARGET]) # T Team

### **MODEL**

### CT Regressor

# regressor = LGBMClassifier(num_leaves=50,
#                           n_estimators=350,
#                           min_data_in_leaf=15,
#                           max_depth=7,
#                           learning_rate=0.055,
#                           feature_fraction=0.8,
#                           bagging_frequency=0.75,
#                           bagging_fraction=0.75)

### T Regressor

regressor = LGBMClassifier(num_leaves=50,
                          n_estimators=350,
                          min_data_in_leaf=15,
                          max_depth=7,
                          learning_rate=0.055,
                          feature_fraction=0.8,
                          bagging_frequency=0.75,
                          bagging_fraction=0.75)

### Models

# ct_model = Pipeline(steps=[('ct_preprocessor', preprocessor_ct),
#                            ('regressor', regressor)])

t_model = Pipeline(steps=[('t_preprocessor', preprocessor_t),
                           ('regressor', regressor)])

# ct_model.fit(X_train, y_train);

t_model.fit(X_train, y_train);

### **CHECK PERFORMANCE**

# ct_model.score(X_test, y_test)

t_model.score(X_test, y_test)

# sk_report = classification_report(
#     digits=6,
#     y_true=y_test, 
#     y_pred=ct_model.predict(X_test))
# print(sk_report)


sk_report = classification_report(
    digits=6,
    y_true=y_test, 
    y_pred=t_model.predict(X_test))
print(sk_report)

### **TRAIN MODEL WITH FULL DATASET**

#### **CT Model. Train & Save**

regressor = LGBMClassifier(num_leaves=50,
                          n_estimators=350,
                          min_data_in_leaf=15,
                          max_depth=7,
                          learning_rate=0.055,
                          feature_fraction=0.8,
                          bagging_frequency=0.75,
                          bagging_fraction=0.75)

ct_model = Pipeline(steps=[('ct_preprocessor', preprocessor_ct),
                           ('regressor', regressor)])

ct_model.fit(ct_df[CT_FEATS], ct_df[TARGET]);

ct_df_pred_nxt_rnd = ct_model.predict(ct_df[CT_FEATS])

dump(ct_model, '../models/ct_nxt_rnd_type.joblib') # Save model

ct_df['ct_nxt_rnd_type_pred'] = ct_df_pred_nxt_rnd

ct_df.to_csv('../data/results/ct_predicted_nxt_rnd.csv', index=False) # Export the CT DataFrame with the prediction column added

#### **T Model. Train & Save**

regressor = LGBMClassifier(num_leaves=50,
                          n_estimators=350,
                          min_data_in_leaf=15,
                          max_depth=7,
                          learning_rate=0.055,
                          feature_fraction=0.8,
                          bagging_frequency=0.75,
                          bagging_fraction=0.75)

t_model = Pipeline(steps=[('t_preprocessor', preprocessor_t),
                           ('regressor', regressor)])

t_model.fit(t_df[T_FEATS], t_df[TARGET]);

t_df_pred_nxt_rnd = t_model.predict(t_df[T_FEATS])

dump(t_model, '../models/t_nxt_rnd_type.joblib') # Save model

t_df['t_nxt_rnd_type_pred'] = t_df_pred_nxt_rnd

t_df.to_csv('../data/results/t_predicted_nxt_rnd.csv', index=False) # Export the T DataFrame with the prediction column added

### **JOINING DATA AND MAKE IT LEGIBLE**

### Load Data

ct_df = pd.read_csv('../data/results/ct_predicted_nxt_rnd.csv')
t_df = pd.read_csv('../data/results/t_predicted_nxt_rnd.csv')

display(ct_df.head())
display(t_df.head())

### Let's decode the round type columns

round_type_dic_decode = {0:'PISTOL_ROUND', 1:'ECO', 2:'MEDIUM', 3:'FULL', 4:'LAST'}

ct_df['round_type'] = ct_df['round_type'].apply(lambda x: round_type_dic_decode[x])
ct_df['nxt_rnd_type'] = ct_df['nxt_rnd_type'].apply(lambda x: round_type_dic_decode[x])
ct_df['ct_nxt_rnd_type_pred'] = ct_df['ct_nxt_rnd_type_pred'].apply(lambda x: round_type_dic_decode[x])

t_df['round_type'] = t_df['round_type'].apply(lambda x: round_type_dic_decode[x])
t_df['nxt_rnd_type'] = t_df['nxt_rnd_type'].apply(lambda x: round_type_dic_decode[x])
t_df['t_nxt_rnd_type_pred'] = t_df['t_nxt_rnd_type_pred'].apply(lambda x: round_type_dic_decode[x])

display(ct_df.head())
display(t_df.head())

ct_df.rename(columns={'round_type': 'ct_round_type', 'nxt_rnd_type':'ct_nxt_rnd_type'}, inplace=True)
t_df.rename(columns={'round_type': 't_round_type', 'nxt_rnd_type':'t_nxt_rnd_type'}, inplace=True)

### The next move is trying to predict which team will win the next round. 

### For this prediction we will need only one DataFrame, so let's join them.

The resulting DataFrame should have these columns:
- file
- round
- wp_ct_val
- wp_t_val
- nade_ct_val
- nade_t_val
- ct_alive
- t_alive
- ct_winner
- bomb_planted
- ct_cons_wins
- t_cons_wins
- ct_val_pred
- t_val_pred
- ct_round_type
- t_round_type
- ct_nxt_rnd_type_pred
- t_nxt_rnd_type_pred

df = pd.concat([ct_df, t_df], axis=1)

columns = ['file','round','wp_ct_val','wp_t_val','nade_ct_val','nade_t_val','ct_alive',
           't_alive','ct_winner','bomb_planted','ct_cons_wins','t_cons_wins','ct_val_pred', 
           't_val_pred', 'ct_round_type', 't_round_type','ct_nxt_rnd_type', 't_nxt_rnd_type',
           'ct_nxt_rnd_type_pred', 't_nxt_rnd_type_pred']

df = df[columns]
df = df.loc[:,~df.columns.duplicated()]
df = df[columns]
df.head()

df.rename(columns={'ct_winner':'prev_ct_winner'}, inplace=True)

#### The column we had <code>'ct_winner'</code> that has been renamed is because it defines the winner of the last round.

#### We need to add a column that defines the winner of the actual round.

#### Let's take this column from the original dataset <code>esea_meta_demos.csv</code>

meta_df = pd.read_csv('../data/csgo/esea_meta_demos.csv')
meta_df_flt = meta_df[['file', 'round', 'winner_side']]

Let's take only the files we took at the beginning, only the games with all the rounds registered

files_complete = [file for file in meta_df_flt['file'].unique() if\
                     meta_df_flt[meta_df_flt['file'] == file]['round'].max() == \
                     len(meta_df_flt[meta_df_flt['file'] == file])]

winner_list = []

for file in files_complete:
    meta_file = meta_df_flt[meta_df_flt['file'] == file]
    winner_list.append(meta_file['winner_side'].to_list())

winner_list = [item for sublist in winner_list for item in sublist] # Flat the list

df['ct_winner'] = winner_list #Add the new column

Reorder the columns to put <code>prev_ct_winner</code> and <code>ct_winner</code> together

df.columns

df = df[['file', 'round', 'wp_ct_val', 'wp_t_val', 'nade_ct_val', 'nade_t_val',
       'ct_alive', 't_alive', 'prev_ct_winner', 'ct_winner', 'bomb_planted', 'ct_cons_wins',
       't_cons_wins', 'ct_val_pred', 't_val_pred', 'ct_round_type',
       't_round_type', 'ct_nxt_rnd_type', 't_nxt_rnd_type',
       'ct_nxt_rnd_type_pred', 't_nxt_rnd_type_pred']]

df.head()

### Encode <code>ct_winner</code> column

winner_dict = {'CounterTerrorist': 1, 'Terrorist':0}

df['ct_winner'] = df['ct_winner'].apply(lambda x: winner_dict[x])

df.head(6)

### **SAVE DATA FOR WINNER PREDICTION**

df.to_csv('../data/processed/5_base_predict_winner.csv', index=False)
