==> notebooks/5_1_prepr_ml_clas_winner_team.ipynb <==
## **Supervised ML classification algorithm to predict the winner team (CT or T)**
## **Encoding**

import pandas as pd

pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', 30)

df = pd.read_csv('../data/processed/5_base_predict_winner.csv')

df.head()

round_type_dic = {'PISTOL_ROUND':0, 'ECO':1, 'MEDIUM':2, 'FULL':3, 'LAST':4}

columns_encoding = ['ct_round_type', 'ct_nxt_rnd_type', 'ct_nxt_rnd_type_pred',
                    't_round_type', 't_nxt_rnd_type', 't_nxt_rnd_type_pred']

for col in columns_encoding:
    df[col] = df[col].apply(lambda x: round_type_dic[x])

df.head()

## **Preprocessing**

The <code>bomb_planted</code> column refers to the previous round. Let's specify this by changing the name to <code>prev_bomb_planted</code> and let's add a new column <code>bomb_planted</code> with **1 if the bomb is planted in the current round and 0 if not**.

We will take this column from the original dataset <code>esea_master_kills_demos.csv</code>

df.rename(columns={'bomb_planted':'prev_bomb_planted'}, inplace=True)

meta_demos = pd.read_csv('../data/csgo/esea_meta_demos.csv')
kills_demos = pd.read_csv('../data/csgo/esea_master_kills_demos.csv') 

We load the dataset <code>esea_meta_demos.csv</code> to take the list of files we are woring with

files_complete = [file for file in meta_demos['file'].unique() if\
                     meta_demos[meta_demos['file'] == file]['round'].max() == \
                     len(meta_demos[meta_demos['file'] == file])]

bomb_planted = []

for file in files_complete:
    kills_file = kills_demos[kills_demos['file'] == file]
    rounds = kills_file['round'].unique()
    for rnd in rounds:
        kills_rnd = kills_file[kills_file['round'] == rnd]
        if True in kills_rnd['is_bomb_planted'].to_list():
            bomb_planted.append(1)
        else:
            bomb_planted.append(0)

print(len(bomb_planted))
print(len(df))

df['bomb_planted'] = bomb_planted

df.columns

columns = ['file', 'round', 'wp_ct_val', 'wp_t_val', 'nade_ct_val', 'nade_t_val',
       'ct_alive', 't_alive', 'prev_ct_winner', 'ct_winner',
       'prev_bomb_planted', 'bomb_planted', 'ct_cons_wins', 't_cons_wins', 'ct_val_pred',
       't_val_pred', 'ct_round_type', 't_round_type', 'ct_nxt_rnd_type',
       't_nxt_rnd_type', 'ct_nxt_rnd_type_pred', 't_nxt_rnd_type_pred']

df = df[columns]
df.head()

### **Create the TARGET** (column to predict)
#### <code>nxt_ct_winner</code>

%%time

files = df['file'].unique()

nxt_ct_winner = []

log = 0

for file in files:

    df_file = df[df['file'] == file] 
    rounds = df_file['round'].unique()
    
    if log%1000 == 0:
        print(f'{log}/12185')    
    log += 1

    for rnd in rounds:
        if rnd == rounds[-1]:
            nxt_ct_winner.append(0)
        else:
            nxt_ct_winner.append(df_file[df_file['round'] == rnd + 1]['bomb_planted'].values[0])
            
df['nxt_ct_winner'] = nxt_ct_winner

df.head()

#### Last step before going into the ml classification algorithm, let's remove the following columns:
- <code>ct_nxt_rnd_type</code>
- <code>t_nxt_rnd_type</code> 	


df.drop(columns=['ct_nxt_rnd_type', 't_nxt_rnd_type'], inplace=True)

df.head()

df.to_csv('../data/processed/5_1_base_predict_winner.csv', index=False)
